requirements.txt
manifold3d>=3.3.2
mapbox_earcut>=2.0.0
fonttools>=4.0.0
markdown>=3.6
Pillow>=10.0.0
pyvista>=0.43.0
rich>=13.7.0
typer>=0.9.0
watchfiles>=0.21.0
pyclipper>=1.3.0
